requirements.txt
streamlit
torch
pillow
pandas
numpy
scikit-learn
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
